requirements.txt
numpy==1.21.5
Pillow==8.4.0
PyYAML==6.0
six==1.16.0
wxPython==4.1.1
